% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associated_eqtl.R
\name{associated_eqtl}
\alias{associated_eqtl}
\title{The known eQTL for snps or genes (eQTL catalogue)}
\usage{
associated_eqtl(snps = c(), genes = c(), p_cutoff = 1e-05, verbose = FALSE)
}
\arguments{
\item{snps}{A list of snp names}

\item{genes}{A list of gene names (ensembl IDs)}

\item{p_cutoff}{The pvalue cutoff for returning associations}

\item{verbose}{Whether to spit out results}
}
\description{
The known eQTL for snps or genes (eQTL catalogue)
}
\references{
https://www.ebi.ac.uk/eqtl/api-docs/
}
