% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associated_publications.R
\name{associated_publications}
\alias{associated_publications}
\title{What publications are available?}
\usage{
associated_publications(
  genes,
  keywords = c("rheumatoid", "inflammatory", "autoimmune"),
  split = "OR",
  long_format = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{genes}{A list of gene names}

\item{keywords}{Words to search with the genes}

\item{split}{search keywords using AND or OR (default is OR)}

\item{long_format}{Logical whether to return one row per publication (long) or one per row per gene (wide)}
}
\description{
What publications are available?
}
