% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associated_snps.R
\name{associated_snps}
\alias{associated_snps}
\title{What snps are known to be associated with gene (GWAS catalogue)}
\usage{
associated_snps(genes = c(), delim = "; ", verbose = FALSE)
}
\arguments{
\item{genes}{A list of gene names (symbol or ensembl)}

\item{delim}{The deliminator for returning traits}

\item{verbose}{Whether to spit out results}
}
\description{
What snps are known to be associated with gene (GWAS catalogue)
}
