% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associated_traits.R
\name{associated_traits}
\alias{associated_traits}
\title{What traits are known to be associated with snps (GWAS catalogue)}
\usage{
associated_traits(snps, delim = "; ", verbose = FALSE)
}
\arguments{
\item{snps}{A list of snp names}

\item{delim}{The deliminator for returning traits}

\item{verbose}{Whether to spit out results}
}
\description{
What traits are known to be associated with snps (GWAS catalogue)
}
\references{
https://www.ebi.ac.uk/gwas/rest/docs/api
}
