% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enriched_pathways.R
\name{enriched_pathways}
\alias{enriched_pathways}
\title{returns pathways from enrichR}
\usage{
enriched_pathways(
  genes,
  drop_terms = c("User", "Enrichr"),
  keep_terms = c(),
  dbs = NULL,
  cutoff = 0.05,
  min_N = 2,
  remove_old = TRUE,
  libraries = c("Transcription", "Pathways"),
  plot_cap = 20,
  plot_pcutoff = 0.01,
  plot_colour = "Combined.Score",
  plot_x = "P.value",
  plot_vline = NA,
  check_for_updates = TRUE
)
}
\arguments{
\item{genes}{List of gene names}

\item{drop_terms}{terms in databases not to search. If NULL only keep_terms
used}

\item{keep_terms}{terms in databases to search. If NULL only all except
drop_terms used}

\item{dbs}{database of enrichR pathways. If NULL all are selected minus
those removed by drop_terms. See listEnrichrDbs().}

\item{cutoff}{pvalue cutoff}

\item{min_N}{minimum number of genes to consider}

\item{remove_old}{Whether to remove legacy libraries ()}

\item{libraries}{Which type of libraries to check for enrichment. Options
include c('Transcription', 'Pathways', 'Ontologies', 'Diseases_Drugs',
'Cell_Types', 'Misc'))}

\item{plot_cap}{Maximum number of terms to plot}

\item{plot_pcutoff}{Pvalue cutoff for plotting}

\item{plot_colour}{Parameter for the colour scaling}

\item{plot_x}{Parameter to plot on the x-axis}

\item{plot_vline}{Position for vertical line}

\item{check_for_updates}{Whether to check for updates in the enrichR database}
}
\description{
returns pathways from enrichR
}
