% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_description.R
\name{full_gene_name}
\alias{full_gene_name}
\title{full gene names}
\usage{
full_gene_name(genes, name_type = "hgnc_symbol")
}
\arguments{
\item{genes}{A list of gene names}

\item{name_type}{What format are the names}
}
\description{
full gene names
}
