% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_description.R
\name{gene_description}
\alias{gene_description}
\title{Rentrez gene descriptions}
\usage{
gene_description(genes)
}
\arguments{
\item{genes}{A list of gene names}
}
\description{
Rentrez gene descriptions
}
