% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_summary.R
\name{gene_summary}
\alias{gene_summary}
\title{wrapper function to find all info about a list of genes}
\usage{
gene_summary(
  genes,
  verbose = TRUE,
  get_gene_types = TRUE,
  get_gene_description = TRUE,
  get_associated_diseases = TRUE,
  get_publications = FALSE,
  gene_types_df = NULL,
  ncbi_retmax = 50,
  disease_cutoff = 0,
  disease_source = "curated",
  disease_api_token = NULL,
  diseases = c("C20", "C05", "C10", "C17"),
  publication_keywords = c("Rheumatoid", "Autoimmune"),
  publication_split = "OR"
)
}
\arguments{
\item{genes}{List of gene names}

\item{verbose}{Whether to split out the progress}

\item{get_gene_types}{Highlight interesting genes?}

\item{get_gene_description}{Whether to find a summary of the gene info}

\item{get_associated_diseases}{Whether to find the associated diseases}

\item{get_publications}{Do you want publication info? (This is time consuming
if you have a lot of genes)}

\item{gene_types_df}{A data frame of annotation types with columns for: the
annotation name, the contains a
character vector for grep command and general info about the gene. If NULL
system.file("extdata","immune_genes.txt", package = "omicAnnotations")}

\item{ncbi_retmax}{The number of NCBI hits to try. If you are struggling to
find info try increasing this, although it will become slower (default=50).}

\item{disease_cutoff}{The DisGeNET Score cutoff}

\item{disease_source}{The source of the GDA (see
https://www.disgenet.org/dbinfo)}

\item{disease_api_token}{The authentication token by DisGenNet (Sign up here:
https://www.disgenet.org/signup/). Then run get_api_key(email, password).}

\item{diseases}{The MeSH disease classes to use
("C20"=Immune System Diseases)}

\item{publication_keywords}{Words to search for publications with the genes}

\item{publication_split}{Whether to use AND or OR when searching the
publication keywords (default is OR)}
}
\description{
wrapper function to find all info about a list of genes
}
