% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_types.R
\name{gene_types}
\alias{gene_types}
\title{genes EMR are interested in}
\usage{
gene_types(genes, types = NULL)
}
\arguments{
\item{genes}{A list of gene names}

\item{types}{A data frame of annotation types with columns for: the
annotation name, the contains a
character vector for grep command and general info about the gene. If NULL
immune_genes are used}
}
\description{
genes EMR are interested in
}
