% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associated_diseases.R
\name{get_api_key}
\alias{get_api_key}
\title{Retrieves the DisGeNET API key for a user}
\usage{
get_api_key(email, password)
}
\arguments{
\item{email}{the email of the user}

\item{password}{the password of the user to connect to DisGeNET}
}
\value{
A string
}
\description{
Given the email and password, retrieves the users API Key
}
\examples{
# Not a real account so will return an error
get_api_key(email="user@gmail.com", password="myspwd")
}
