% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex_eqtl.R
\name{gtex_heatmap}
\alias{gtex_heatmap}
\title{gtex eqtl heatmap}
\usage{
gtex_heatmap(
  df,
  colour_low = "blue",
  colour_high = "white",
  grid_colour = NULL,
  p_cutoff = 0.05,
  ...
)
}
\arguments{
\item{df}{A list of gene names in format genecode IDs}

\item{colour_low}{Colour for most significant}

\item{colour_high}{Colour for least significant}

\item{grid_colour}{Colour for heatmap grid}

\item{...}{}
}
\description{
gtex eqtl heatmap
}
