% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_genes.R
\docType{data}
\name{immune_genes}
\alias{immune_genes}
\title{List of gene name structures and types}
\format{
An object of class \code{list} of length 51.
}
\usage{
immune_genes
}
\description{
List of gene name structures and types
}
\keyword{datasets}
