% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lymphoid_pathways.R
\docType{data}
\name{lymphoid_pathways}
\alias{lymphoid_pathways}
\title{Enriched pathways for lymphoid genes}
\format{
An object of class \code{list} of length 3.
}
\usage{
lymphoid_pathways
}
\description{
list of: enriched pathways, enriched genes, plot of enrichment
}
\keyword{datasets}
