% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associated_diseases.R
\name{associated_diseases}
\alias{associated_diseases}
\title{What diseases are known to be associated with genes}
\usage{
associated_diseases(
  genes,
  cutoff = 0,
  disease_classes = c("C20", "C05", "C10", "C17"),
  delim = "; ",
  verbose = TRUE
)
}
\arguments{
\item{genes}{A list of gene names}

\item{cutoff}{The DisGeNET Score cutoff}

\item{disease_classes}{The MeSH disease classes to use
("C20"=Immune System Diseases)}

\item{delim}{The deliminator for returning traits}

\item{verbose}{Whether to spit out results}
}
\description{
What diseases are known to be associated with genes
}
\references{
https://www.disgenet.org/api/#/GDA/gdaByGene
}
