% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_description.R
\name{gene_description}
\alias{gene_description}
\title{Rentrez gene descriptions}
\usage{
gene_description(genes, retmax = 50)
}
\arguments{
\item{genes}{A list of gene names}

\item{retmax}{The number of NCBI hits to try. If you are struggling to find
info try increasing this, although it will become slower (default=50).}
}
\description{
Rentrez gene descriptions
}
