% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_summary.R
\name{gene_summary}
\alias{gene_summary}
\title{wrapper function to find all info about a list of genes}
\usage{
gene_summary(
  genes,
  verbose = TRUE,
  gene_types = TRUE,
  gene_description = TRUE,
  associated_diseases = TRUE,
  publications = FALSE,
  gene_types_list = NULL,
  disease_cutoff = 0,
  diseases = c("C20", "C05", "C10", "C17"),
  publication_keywords = c("Rheumatoid", "Autoimmune"),
  publication_split = "OR"
)
}
\arguments{
\item{genes}{List of gene names}

\item{verbose}{Whether to split out the progress}

\item{gene_types}{Highlight interesting genes?}

\item{associated_diseases}{Whether to find the associated diseases}

\item{publications}{Do you want pulication info? (This is time consuming
if you have a lot of genes)}

\item{disease_cutoff}{The DisGeNET Score cutoff}

\item{diseases}{The MeSH disease classes to use
("C20"=Immune System Diseases)}

\item{publication_keywords}{Words to search for publications with the genes}

\item{publication_split}{Whether to use AND or OR when searching the
publication keywords (default is OR)}

\item{gene_summaries}{Whether to find a summary of the gene info}
}
\description{
wrapper function to find all info about a list of genes
}
